Separating the coredata from the delegate
